Split oversized backend and E2E test files
