--- apps/server/tests/hygiene/test_domain_architecture.py
@@ -1,1135 +1,0 @@
-"""Architecture guardrails: domain-first pipeline boundaries.
-
-Ensures that core analysis modules operate on domain ``Finding`` objects
-(not raw dicts) and that ``TestRun`` is the canonical post-analysis
-aggregate.  Prevents regression to the previous payload-first flow.
-"""
-
-from __future__ import annotations
-
-import pytest
-
-
-def test_test_run_is_frozen_dataclass() -> None:
-    """``TestRun`` must be a frozen dataclass."""
-    import dataclasses
-
-    from vibesensor.domain import RunCapture, TestRun
-
-    assert dataclasses.is_dataclass(TestRun)
-    r = TestRun(
-        capture=RunCapture(run_id="test-123"),
-        findings=(),
-        top_causes=(),
-    )
-    with pytest.raises(AttributeError):
-        r.findings = ()
-
-
-# ── TestRun provides domain queries ──────────────────────────────────────
-
-
-def test_test_run_provides_finding_queries() -> None:
-    """``TestRun`` must own finding classification queries."""
-    from vibesensor.domain import Finding, RunCapture, TestRun
-
-    diag = Finding(finding_id="F001", confidence=0.75, suspected_source="wheel/tire")
-    ref = Finding(finding_id="REF_SPEED", confidence=1.0, suspected_source="unknown")
-    info = Finding(finding_id="F002", confidence=0.10, severity="info", suspected_source="unknown")
-
-    result = TestRun(
-        capture=RunCapture(run_id="test-123"),
-        findings=(ref, diag, info),
-        top_causes=(diag,),
-    )
-
-    assert result.primary_finding == diag
-    assert result.diagnostic_findings == (diag,)
-    assert result.non_reference_findings == (diag, info)
-
-
-def test_test_run_effective_top_causes() -> None:
-    """``effective_top_causes()`` mirrors diagnosis_candidates logic."""
-    from vibesensor.domain import Finding, RunCapture, TestRun
-
-    actionable = Finding(finding_id="F001", confidence=0.75, suspected_source="wheel/tire")
-    result = TestRun(
-        capture=RunCapture(run_id="test"),
-        findings=(actionable,),
-        top_causes=(actionable,),
-    )
-    effective = result.effective_top_causes()
-    assert actionable in effective
-
-
-# ── finalize_findings returns domain objects ──────────────────────────────
-
-
-def test_finalize_findings_returns_domain_findings() -> None:
-    """``finalize_findings`` must return domain ``Finding`` objects."""
-    from vibesensor.domain import Finding
-    from vibesensor.use_cases.diagnostics.findings import finalize_findings
-
-    domain_findings = finalize_findings(
-        [
-            Finding(finding_id="F_ORDER", confidence=0.7, suspected_source="wheel/tire"),
-        ]
-    )
-    assert len(domain_findings) == 1
-    assert isinstance(domain_findings[0], Finding)
-    assert domain_findings[0].finding_id == "F001"
-
-
-# ── select_top_causes returns domain objects ─────────────────────────────
-
-
-def test_select_top_causes_returns_domain_findings() -> None:
-    """``select_top_causes`` must return domain ``Finding`` objects."""
-    from test_support.findings import make_finding_payload
-
-    from vibesensor.domain import Finding
-    from vibesensor.shared.boundaries.finding import finding_from_payload
-    from vibesensor.use_cases.diagnostics.top_cause_selection import select_top_causes
-
-    findings = tuple(
-        finding_from_payload(f)
-        for f in [make_finding_payload(confidence=0.80, suspected_source="wheel/tire")]
-    )
-    domain_findings = select_top_causes(findings)
-    assert len(domain_findings) == 1
-    assert isinstance(domain_findings[0], Finding)
-
-
-# ── RunAnalysis.summarize() produces TestRun ─────────────────────────────
-
-
-def test_run_analysis_produces_test_run() -> None:
-    """``RunAnalysis.summarize()`` must populate ``test_run``."""
-    from vibesensor.domain import TestRun
-    from vibesensor.use_cases.diagnostics.summary_builder import RunAnalysis
-
-    metadata = {"run_id": "test-guard", "car_type": "sedan"}
-    samples = [
-        {
-            "t_s": float(i),
-            "accel_x": 0.01,
-            "accel_y": 0.01,
-            "accel_z": 1.0,
-            "accel_mag": 1.0,
-            "speed_kmh": 80.0,
-            "vibration_strength_db": 5.0,
-        }
-        for i in range(30)
-    ]
-    analysis = RunAnalysis(metadata, samples)
-    result = analysis.summarize()
-    assert analysis.test_run is not None
-    assert isinstance(analysis.test_run, TestRun)
-    assert analysis.test_run.run_id == result.summary["run_id"]
-    assert len(analysis.test_run.findings) == len(result.summary["findings"])
-
-
-# ── TestRun.has_relevant_reference_gap() ─────────────────────────────────
-
-
-def test_test_run_reference_gap_detection() -> None:
-    """``has_relevant_reference_gap()`` detects source-relevant gaps."""
-    from vibesensor.domain import Finding, RunCapture, TestRun, VibrationSource
-
-    ref_speed = Finding(finding_id="REF_SPEED", suspected_source="unknown")
-    ref_wheel = Finding(finding_id="REF_WHEEL", suspected_source="unknown")
-    ref_engine = Finding(finding_id="REF_ENGINE", suspected_source="unknown")
-    diag = Finding(finding_id="F001", confidence=0.80, suspected_source="wheel/tire")
-
-    assert ref_speed.is_reference
-    assert ref_wheel.is_reference
-    assert ref_engine.is_reference
-    assert not diag.is_reference
-
-    def _make(findings: tuple[Finding, ...]) -> TestRun:
-        return TestRun(
-            capture=RunCapture(run_id="test"),
-            findings=findings + (diag,),
-            top_causes=(diag,),
-        )
-
-    result = _make((ref_speed,))
-    assert result.has_relevant_reference_gap(VibrationSource.WHEEL_TIRE)
-    assert result.has_relevant_reference_gap(VibrationSource.ENGINE)
-
-    result2 = _make((ref_wheel,))
-    assert result2.has_relevant_reference_gap(VibrationSource.WHEEL_TIRE)
-    assert result2.has_relevant_reference_gap(VibrationSource.DRIVELINE)
-    assert not result2.has_relevant_reference_gap(VibrationSource.ENGINE)
-
-    result3 = _make((ref_engine,))
-    assert result3.has_relevant_reference_gap(VibrationSource.ENGINE)
-    assert not result3.has_relevant_reference_gap(VibrationSource.WHEEL_TIRE)
-
-
-# ── TestRun.top_strength_db() ────────────────────────────────────────────
-
-
-def test_test_run_top_strength_db() -> None:
-    """``top_strength_db()`` finds the best strength from findings."""
-    from vibesensor.domain import Finding, RunCapture, TestRun
-
-    f1 = Finding(
-        finding_id="F001",
-        confidence=0.80,
-        suspected_source="wheel/tire",
-        vibration_strength_db=12.5,
-    )
-    f2 = Finding(
-        finding_id="F002",
-        confidence=0.60,
-        suspected_source="engine",
-        vibration_strength_db=8.0,
-    )
-    result = TestRun(
-        capture=RunCapture(run_id="test"),
-        findings=(f1, f2),
-        top_causes=(f1,),
-    )
-    assert result.top_strength_db() == 12.5
-
-    # No strength → None
-    f3 = Finding(finding_id="F003", confidence=0.50, suspected_source="engine")
-    result2 = TestRun(
-        capture=RunCapture(run_id="test"),
-        findings=(f3,),
-        top_causes=(f3,),
-    )
-    assert result2.top_strength_db() is None
-
-
-# ── Report mapping context builds domain aggregate ───────────────────────
-
-
-def test_report_mapping_context_has_domain_aggregate() -> None:
-    """``prepare_report_mapping_context`` must build a domain aggregate."""
-    from test_support.findings import make_finding_payload
-
-    from vibesensor.adapters.pdf.mapping import prepare_report_mapping_context
-    from vibesensor.domain import TestRun
-
-    summary = {
-        "run_id": "test-context",
-        "findings": [make_finding_payload(finding_id="F001")],
-        "top_causes": [make_finding_payload(finding_id="F001")],
-        "lang": "en",
-        "metadata": {},
-        "report_date": "",
-        "record_length": "",
-        "start_time_utc": "",
-        "end_time_utc": "",
-        "sensor_locations": [],
-        "sensor_locations_connected_throughout": [],
-        "most_likely_origin": {},
-        "run_suitability": [],
-    }
-    context = prepare_report_mapping_context(summary)
-    assert context.domain_aggregate is not None
-    assert isinstance(context.domain_aggregate, TestRun)
-    assert len(context.domain_aggregate.findings) == 1
-
-
-# ── Finding owns confidence presentation ─────────────────────────────────
-
-
-def test_finding_owns_confidence_label() -> None:
-    """Finding must own confidence-tier classification (label, tone, pct)."""
-    from vibesensor.domain import Finding
-
-    high = Finding(finding_id="F001", confidence=0.80, suspected_source="wheel/tire")
-    label_key, tone, pct_text = high.confidence_label()
-    assert label_key == "CONFIDENCE_HIGH"
-    assert tone == "success"
-    assert pct_text == "80%"
-
-    medium = Finding(finding_id="F002", confidence=0.55, suspected_source="engine")
-    label_key, tone, _ = medium.confidence_label()
-    assert label_key == "CONFIDENCE_MEDIUM"
-    assert tone == "warn"
-
-    low = Finding(finding_id="F003", confidence=0.20, suspected_source="unknown")
-    label_key, tone, _ = low.confidence_label()
-    assert label_key == "CONFIDENCE_LOW"
-    assert tone == "neutral"
-
-
-def test_finding_confidence_negligible_strength_downgrade() -> None:
-    """Finding with negligible strength should downgrade HIGH → MEDIUM."""
-    from vibesensor.domain import Finding
-
-    high = Finding(finding_id="F001", confidence=0.80, suspected_source="wheel/tire")
-    label_key, tone, _ = high.confidence_label(strength_band_key="negligible")
-    assert label_key == "CONFIDENCE_MEDIUM"
-    assert tone == "warn"
-
-
-def test_classify_confidence_is_domain_owned() -> None:
-    """Finding.classify_confidence is the canonical source of truth for confidence presentation."""
-    from vibesensor.domain import Finding
-
-    for conf in (0.80, 0.55, 0.20, 0.0):
-        result = Finding.classify_confidence(conf)
-        assert len(result) == 3
-        assert all(isinstance(v, str) for v in result)
-
-
-# ── TestRun owns primary source/location queries ─────────────────────────
-
-
-def test_test_run_primary_source_and_location() -> None:
-    """``primary_source`` and ``primary_location`` are domain queries."""
-    from vibesensor.domain import Finding, RunCapture, TestRun, VibrationSource
-
-    f = Finding(
-        finding_id="F001",
-        confidence=0.80,
-        suspected_source="wheel/tire",
-        strongest_location="Left Front",
-    )
-    result = TestRun(
-        capture=RunCapture(run_id="test"),
-        findings=(f,),
-        top_causes=(f,),
-    )
-    assert result.primary_source == VibrationSource.WHEEL_TIRE
-    assert result.primary_location == "Left Front"
-
-    # No findings → None
-    empty = TestRun(
-        capture=RunCapture(run_id="test"),
-        findings=(),
-        top_causes=(),
-    )
-    assert empty.primary_source is None
-    assert empty.primary_location is None
-
-
-# ── Domain objects are immutable ──────────────────────────────────────────
-
-
-def test_finding_is_frozen_dataclass() -> None:
-    """Finding must be a frozen dataclass — mutation raises."""
-    import dataclasses
-
-    from vibesensor.domain import Finding
-
-    assert dataclasses.is_dataclass(Finding)
-    f = Finding(finding_id="F001", confidence=0.80, suspected_source="wheel/tire")
-    with pytest.raises(AttributeError):
-        f.confidence = 0.50
-
-
-def test_finding_tuples_are_immutable() -> None:
-    """TestRun findings must be tuples (not mutable lists)."""
-    from vibesensor.domain import Finding, RunCapture, TestRun
-
-    f = Finding(finding_id="F001", confidence=0.80, suspected_source="wheel/tire")
-    result = TestRun(
-        capture=RunCapture(run_id="test"),
-        findings=(f,),
-        top_causes=(f,),
-    )
-    assert isinstance(result.findings, tuple)
-    assert isinstance(result.top_causes, tuple)
-
-
-# ── Report mapping uses domain aggregate ─────────────────────────────────
-
-
-def test_build_system_cards_uses_domain_findings() -> None:
-    """build_system_cards must read confidence tone from domain, not dict."""
-    from vibesensor.adapters.pdf.mapping import (
-        PrimaryCandidateContext,
-        ReportMappingContext,
-        build_system_cards,
-    )
-    from vibesensor.domain import Finding, RunCapture, TestRun
-    from vibesensor.report_i18n import tr
-
-    lang = "en"
-    domain_f = Finding(
-        finding_id="F001",
-        confidence=0.80,
-        suspected_source="wheel/tire",
-        strongest_location="Left Front",
-    )
-    aggregate = TestRun(
-        capture=RunCapture(run_id="test"),
-        findings=(domain_f,),
-        top_causes=(domain_f,),
-    )
-    # Build a context with the aggregate (payloads no longer stored on context)
-    context = ReportMappingContext(
-        car_name=None,
-        car_type=None,
-        date_str="",
-        origin={},
-        origin_location="",
-        sensor_locations_active=[],
-        duration_text=None,
-        start_time_utc=None,
-        end_time_utc=None,
-        sample_rate_hz=None,
-        tire_spec_text=None,
-        sample_count=0,
-        sensor_model=None,
-        firmware_version=None,
-        domain_aggregate=aggregate,
-    )
-    primary = PrimaryCandidateContext(
-        primary_candidate=domain_f,
-        primary_source="wheel/tire",
-        primary_system="Wheel/Tire",
-        primary_location="Left Front",
-        primary_speed="80-90 km/h",
-        confidence=0.80,
-        sensor_count=2,
-        weak_spatial=False,
-        has_reference_gaps=False,
-        strength_db=12.0,
-        strength_text="Moderate (12.0 dB)",
-        strength_band_key="moderate",
-        certainty_key="high",
-        certainty_label_text="High",
-        certainty_pct="80%",
-        certainty_reason="Consistent order-tracking match",
-        tier="C",
-    )
-    cards = build_system_cards(context, primary, lang, lambda key, **kw: tr(lang, key, **kw))
-    assert len(cards) == 1
-    # The tone should come from domain Finding, not the payload's "WRONG_TONE"
-    assert cards[0].tone != "WRONG_TONE"
-    assert cards[0].tone == "success"  # HIGH confidence → success
-
-
-# ── Pipeline produces domain aggregate for report ─────────────────────────
-
-
-def test_map_summary_produces_report_with_domain_findings() -> None:
-    """map_summary must produce report data using domain-first pipeline."""
-    from test_support.findings import make_finding_payload
-
-    from vibesensor.adapters.pdf.mapping import map_summary
-
-    summary = {
-        "run_id": "test-map",
-        "file_name": "test.csv",
-        "rows": 100,
-        "duration_s": 60.0,
-        "lang": "en",
-        "metadata": {},
-        "report_date": "",
-        "record_length": "",
-        "start_time_utc": "",
-        "end_time_utc": "",
-        "warnings": [],
-        "sensor_locations": [],
-        "sensor_locations_connected_throughout": [],
-        "sensor_intensity_by_location": [],
-        "most_likely_origin": {},
-        "run_suitability": [],
-        "plots": {},
-        "test_plan": [],
-        "findings": [make_finding_payload(finding_id="F001", confidence=0.80)],
-        "top_causes": [make_finding_payload(finding_id="F001", confidence=0.80)],
-        "sensor_count_used": 2,
-    }
-    template = map_summary(summary)
-    assert template.run_id == "test-map"
-
-
-# ── Domain modules do not import boundary payload types ──────────────────
-
-
-# ── Domain value objects are exported and frozen ────────────────────────
-
-
-@pytest.mark.parametrize(
-    "name",
-    [
-        "FindingEvidence",
-        "LocationHotspot",
-        "ConfidenceAssessment",
-        "SpeedProfile",
-        "RunSuitability",
-        "SuitabilityCheck",
-    ],
-)
-def test_domain_value_objects_are_exported(name: str) -> None:
-    """New domain value objects must be importable from ``vibesensor.domain``."""
-    import importlib
-
-    mod = importlib.import_module("vibesensor.domain")
-    assert hasattr(mod, name), f"{name} must be exported from vibesensor.domain"
-
-
-@pytest.mark.parametrize(
-    "name",
-    [
-        "FindingEvidence",
-        "LocationHotspot",
-        "ConfidenceAssessment",
-        "SpeedProfile",
-        "RunSuitability",
-        "SuitabilityCheck",
-    ],
-)
-def test_domain_value_objects_are_frozen_dataclasses(name: str) -> None:
-    """Domain value objects must be frozen dataclasses."""
-    import dataclasses
-    import importlib
-
-    mod = importlib.import_module("vibesensor.domain")
-    cls = getattr(mod, name)
-    assert dataclasses.is_dataclass(cls), f"{name} must be a dataclass"
-
-
-def test_test_run_from_summary_populates_speed_profile() -> None:
-    """test_run_from_summary extracts SpeedProfile when speed_stats is present."""
-    from vibesensor.shared.boundaries.diagnostic_case import test_run_from_summary
-
-    summary = {
-        "run_id": "test-123",
-        "findings": [],
-        "top_causes": [],
-        "speed_stats": {
-            "min_kmh": 30.0,
-            "max_kmh": 90.0,
-            "steady_speed": True,
-        },
-    }
-    result = test_run_from_summary(summary)
-    assert result.speed_profile is not None, "test_run_from_summary must populate speed_profile"
-    assert result.speed_profile.steady_speed
-
-
-def test_test_run_from_summary_populates_suitability() -> None:
-    """test_run_from_summary extracts RunSuitability when run_suitability is present."""
-    from vibesensor.shared.boundaries.diagnostic_case import test_run_from_summary
-
-    summary = {
-        "run_id": "test-123",
-        "findings": [],
-        "top_causes": [],
-        "run_suitability": [
-            {"check_key": "speed", "state": "pass", "explanation": "OK"},
-        ],
-    }
-    result = test_run_from_summary(summary)
-    assert result.suitability is not None, "test_run_from_summary must populate suitability"
-    assert result.suitability.is_usable
-
-
-def test_confidence_assessment_tier_matches_domain_finding() -> None:
-    """ConfidenceAssessment.tier must be consistent with Finding.classify_confidence."""
-    from vibesensor.domain import ConfidenceAssessment, Finding
-
-    for conf in [0.1, 0.3, 0.5, 0.7, 0.9]:
-        label_key, _tone, _pct = Finding.classify_confidence(conf)
-        ca = ConfidenceAssessment.assess(conf)
-        assert ca.label_key == label_key, (
-            f"ConfidenceAssessment.assess({conf}).label_key must match "
-            f"Finding.classify_confidence({conf})[0]"
-        )
-
-
-@pytest.mark.parametrize(
-    "name",
-    [
-        "ConfigurationSnapshot",
-        "DiagnosticCase",
-        "DrivingSegment",
-        "RecommendedAction",
-        "Signature",
-        "Symptom",
-        "TestPlan",
-        "TestRun",
-        "VibrationOrigin",
-    ],
-)
-def test_new_domain_objects_are_exported(name: str) -> None:
-    import importlib
-
-    mod = importlib.import_module("vibesensor.domain")
-    assert hasattr(mod, name), f"{name} must be exported from vibesensor.domain"
-
-
-def test_run_analysis_builds_test_run_and_diagnostic_case() -> None:
-    from vibesensor.domain import DiagnosticCase, TestRun
-    from vibesensor.use_cases.diagnostics.summary_builder import RunAnalysis
-
-    metadata = {
-        "run_id": "domain-case-guard",
-        "car_name": "Guard Car",
-        "car_type": "sedan",
-        "language": "en",
-    }
-    samples = [
-        {
-            "t_s": float(i),
-            "accel_x": 0.01,
-            "accel_y": 0.01,
-            "accel_z": 1.0,
-            "accel_mag": 1.0,
-            "speed_kmh": 80.0,
-            "vibration_strength_db": 5.0,
-        }
-        for i in range(30)
-    ]
-    analysis = RunAnalysis(metadata, samples)
-    result = analysis.summarize()
-
-    assert analysis.test_run is not None
-    assert result.diagnostic_case is not None
-    assert isinstance(analysis.test_run, TestRun)
-    assert isinstance(result.diagnostic_case, DiagnosticCase)
-    assert result.diagnostic_case.primary_run is not None
-    assert result.diagnostic_case.primary_run.run_id == analysis.test_run.run_id
-
-
-def test_boundary_decoder_builds_diagnostic_case_from_summary() -> None:
-    from test_support.findings import make_finding_payload
-
-    from vibesensor.shared.boundaries.diagnostic_case import diagnostic_case_from_summary
-
-    summary = {
-        "case_id": "summary-case-guard-id",
-        "run_id": "summary-case-guard",
-        "metadata": {"car_name": "Guard Car", "car_type": "sedan"},
-        "findings": [make_finding_payload(finding_id="F001", confidence=0.80)],
-        "top_causes": [make_finding_payload(finding_id="F001", confidence=0.80)],
-        "test_plan": [
-            {
-                "action_id": "check-wheel",
-                "what": {"_i18n_key": "ACTION_WHEEL_BALANCE_WHAT"},
-                "why": {"_i18n_key": "ACTION_WHEEL_BALANCE_WHY"},
-            }
-        ],
-    }
-    diagnostic_case = diagnostic_case_from_summary(summary)
-    assert diagnostic_case.case_id == "summary-case-guard-id"
-    assert diagnostic_case.test_runs
-    assert diagnostic_case.primary_run is not None
-
-
-# ── Finding boundary separation ──────────────────────────────────────────
-
-
-def test_finding_has_no_from_payload_method() -> None:
-    """Finding domain class should not own payload decode logic."""
-    from vibesensor.domain import Finding
-
-    assert not hasattr(Finding, "from_payload"), (
-        "Finding.from_payload should live in boundaries/finding.py, not on the domain class"
-    )
-
-
-# ── AnalysisResult coordinator isolation ─────────────────────────────────
-
-
-# ── T6 planning service and file rename guardrails ────────────────────────────
-
-
-def test_finding_projector_in_finding_boundary_module() -> None:
-    """Finding payload projector should live in boundaries/finding.py."""
-    from vibesensor.shared.boundaries.finding import finding_payload_from_domain
-
-    assert callable(finding_payload_from_domain)
-
-
-# ── T7.27: Report mapping rendering-boundary guardrails ──────────────────
-
-
-def test_report_mapping_business_functions_use_domain_objects() -> None:
-    """Key business-decision functions derive values from the domain aggregate.
-
-    When ``prepare_report_mapping_context`` produces a domain aggregate,
-    ``resolve_primary_report_candidate`` must derive primary source,
-    strength, and reference-gap status from domain objects — not from
-    raw payload dict traversal.
-    """
-    from test_support.findings import make_finding_payload
-
-    from vibesensor.adapters.pdf.mapping import (
-        prepare_report_mapping_context,
-        resolve_primary_report_candidate,
-    )
-    from vibesensor.domain import TestRun, VibrationSource
-    from vibesensor.report_i18n import tr
-
-    lang = "en"
-    finding = make_finding_payload(
-        finding_id="F001",
-        confidence=0.80,
-        suspected_source="wheel_tire",
-    )
-    summary = {
-        "run_id": "guard-biz",
-        "findings": [finding],
-        "top_causes": [finding],
-        "lang": lang,
-        "metadata": {},
-        "report_date": "",
-        "record_length": "",
-        "start_time_utc": "",
-        "end_time_utc": "",
-        "sensor_locations": [],
-        "sensor_locations_connected_throughout": [],
-        "most_likely_origin": {},
-        "run_suitability": [],
-    }
-
-    context = prepare_report_mapping_context(summary)
-    assert context.domain_aggregate is not None
-    assert isinstance(context.domain_aggregate, TestRun)
-
-    primary = resolve_primary_report_candidate(
-        context=context,
-        sensor_intensity=[],
-        tr=lambda key, **kw: tr(lang, key, **kw),
-        lang=lang,
-    )
-
-    # primary_source must be a VibrationSource enum (domain-first derivation)
-    assert isinstance(primary.primary_source, VibrationSource), (
-        "primary_source must be a VibrationSource enum when domain aggregate is present"
-    )
-
-
-# ── T9.1-T9.6: Workstream 8 architecture guardrails ──────────────────
-
-
-def test_build_run_suitability_checks_does_not_exist() -> None:
-    """``build_run_suitability_checks`` must not exist in summary_builder.
-
-    It was deleted in Workstream 2 (T3.21).  ``RunSuitability.evaluate()``
-    is the canonical owner of suitability evaluation.
-    """
-    from vibesensor.use_cases.diagnostics import summary_builder
-
-    assert not hasattr(summary_builder, "build_run_suitability_checks"), (
-        "build_run_suitability_checks was deleted; use RunSuitability.evaluate()"
-    )
-
-
-def test_project_analysis_summary_has_no_legacy_summary_version_fast_path() -> None:
-    """project_analysis_summary must always reconstruct through TestRun.
-
-    The legacy `_summary_version == 2` bypass was removed. Historical
-    summaries are reconstructed the same way as current ones so the
-    boundary has a single canonical behavior.
-    """
-    from tests._paths import SERVER_ROOT
-
-    source = (
-        SERVER_ROOT / "vibesensor" / "shared" / "boundaries" / "diagnostic_case.py"
-    ).read_text()
-    assert 'analysis.get("_summary_version") == 2' not in source
-
-
-def test_summary_builder_does_not_define_case_context_wrappers() -> None:
-    """summary_builder must not own duplicate Car/Symptom metadata decoders."""
-    from tests._paths import SERVER_ROOT
-
-    source = (
-        SERVER_ROOT / "vibesensor" / "use_cases" / "diagnostics" / "summary_builder.py"
-    ).read_text()
-    assert "def build_domain_car" not in source
-    assert "def build_domain_symptoms" not in source
-    assert "RunSuitabilityCheck" not in source
-
-
-def test_history_backend_types_do_not_export_history_run_payload() -> None:
-    """History record typing must stay local to history workflows.
-
-    `HistoryRunPayload` made persistence dicts look like a general backend
-    business contract. The only supported alias is the history-local
-    `HistoryRecord` in use_cases/history/helpers.py.
-    """
-    from tests._paths import SERVER_ROOT
-
-    backend_types_source = (
-        SERVER_ROOT / "vibesensor" / "shared" / "types" / "backend_types.py"
-    ).read_text()
-    history_helpers_source = (
-        SERVER_ROOT / "vibesensor" / "use_cases" / "history" / "helpers.py"
-    ).read_text()
-    assert "HistoryRunPayload" not in backend_types_source
-    assert "class HistoryRecord" in history_helpers_source
-
-
-def test_speed_profile_used_by_confidence_assessment() -> None:
-    """ConfidenceAssessment must be the owner of confidence reasoning.
-
-    ``certainty_label()`` was deleted; ``ConfidenceAssessment.assess()``
-    is the single source of truth for confidence assessment. Report mapping
-    uses ``ConfidenceAssessment.tier`` for layout gating.
-    """
-    from tests._paths import SERVER_ROOT
-
-    strength_labels_path = SERVER_ROOT / "vibesensor" / "adapters" / "pdf" / "presentation.py"
-    source = strength_labels_path.read_text()
-    assert "certainty_label" not in source, (
-        "certainty_label was deleted; ConfidenceAssessment.assess() is the replacement"
-    )
-
-
-# ── T14: Architecture guardrail regression tests ─────────────────────────
-
-
-def test_f_order_finding_id_normalization() -> None:
-    """``finalize_findings`` normalises arbitrary IDs to sequential ``F###``.
-
-    ``F_ORDER`` is an internal working ID from order analysis.
-    ``finalize_findings`` replaces all non-reference IDs with stable
-    sequential ``F001``, ``F002``, … so that report and history
-    consumers see deterministic identifiers.  The normalization to
-    ``F001`` is correct behavior, not a test bug.
-    """
-    from vibesensor.domain import Finding
-    from vibesensor.use_cases.diagnostics.findings import finalize_findings
-
-    domain_findings = finalize_findings(
-        [
-            Finding(finding_id="F_ORDER", confidence=0.7, suspected_source="wheel/tire"),
-            Finding(finding_id="F_PERSISTENT", confidence=0.4, suspected_source="engine"),
-        ]
-    )
-    # Both get sequential F### IDs regardless of their input names
-    assert domain_findings[0].finding_id == "F001"
-    assert domain_findings[1].finding_id == "F002"
-    assert all(isinstance(f, Finding) for f in domain_findings)
-
-    # Reference findings keep their original IDs
-    domain_ref = finalize_findings(
-        [
-            Finding(finding_id="REF_SPEED", confidence=None, suspected_source="unknown"),
-            Finding(finding_id="F_ORDER", confidence=0.7, suspected_source="wheel/tire"),
-        ]
-    )
-    assert domain_ref[0].finding_id == "REF_SPEED"
-    assert domain_ref[1].finding_id == "F001"
-
-
-def test_next_steps_domain_path_is_primary() -> None:
-    """``build_next_steps_from_summary`` must use domain aggregate only.
-
-    The function must check ``aggregate.recommended_actions`` as its
-    sole source for next steps.  No payload fallback loop should exist.
-    Prevents regression of T12.
-    """
-    from tests._paths import SERVER_ROOT
-
-    sections_path = SERVER_ROOT / "vibesensor" / "adapters" / "pdf" / "report_sections.py"
-    source = sections_path.read_text()
-
-    # Find the function body
-    func_start = source.find("def build_next_steps_from_summary(")
-    assert func_start != -1, "build_next_steps_from_summary not found in report_sections.py"
-
-    # Find end of function (next top-level def or end of file)
-    next_def = source.find("\ndef ", func_start + 1)
-    func_body = source[func_start : next_def if next_def != -1 else len(source)]
-
-    # The domain aggregate if-guard must exist
-    domain_guard = func_body.find("if aggregate is not None and aggregate.recommended_actions")
-    assert domain_guard != -1, (
-        "build_next_steps_from_summary must check aggregate.recommended_actions"
-    )
-    # No payload fallback loop should remain
-    payload_loop = func_body.find("for step in summary_steps")
-    assert payload_loop == -1, (
-        "build_next_steps_from_summary must not have a payload fallback loop — domain-only"
-    )
-
-
-def test_fallback_payload_functions_removed() -> None:
-    """``top_strength_values`` and ``has_relevant_reference_gap`` have been
-    removed — the domain aggregate is always available."""
-    from vibesensor.adapters.pdf import mapping
-
-    assert not hasattr(mapping, "top_strength_values")
-    assert not hasattr(mapping, "has_relevant_reference_gap")
-
-
-# ── TODO-20: Layer import guardrails ─────────────────────────────────────
-
-
-# ── TODO-22: Canonical domain graph structural verification ──────────────
-
-
-def test_canonical_domain_graph_relationships() -> None:
-    """Verify all canonical domain graph relationships exist as typed fields."""
-    import dataclasses
-
-    from vibesensor.domain import (
-        Car,
-        DiagnosticCase,
-        DrivingSegment,
-        Finding,
-        Measurement,
-        RunCapture,
-        RunSetup,
-        Sensor,
-        SensorPlacement,
-        Signature,
-        SpeedSource,
-        TestRun,
-    )
-
-    def field_type(cls: type, name: str) -> type:
-        """Return the raw annotation for a dataclass field."""
-        hints = {f.name: f for f in dataclasses.fields(cls)}
-        assert name in hints, f"{cls.__name__} missing field {name}"
-        return hints[name]
-
-    # DiagnosticCase
-    field_type(DiagnosticCase, "car")  # Car | None
-    field_type(DiagnosticCase, "test_runs")  # tuple[TestRun, ...]
-
-    # TestRun
-    field_type(TestRun, "capture")  # RunCapture
-    field_type(TestRun, "findings")  # tuple[Finding, ...]
-    field_type(TestRun, "driving_segments")  # tuple[DrivingSegment, ...]
-
-    # RunCapture
-    field_type(RunCapture, "run_id")  # str (not a Run object — known deviation)
-    field_type(RunCapture, "setup")  # RunSetup
-    field_type(RunCapture, "measurements")  # tuple[Measurement, ...]
-    assert not any(f.name == "run" for f in dataclasses.fields(RunCapture)), (
-        "RunCapture must not hold a Run object reference (uses run_id: str)"
-    )
-
-    # RunSetup
-    field_type(RunSetup, "sensors")  # tuple[Sensor, ...]
-    field_type(RunSetup, "speed_source")  # SpeedSource
-
-    # Sensor
-    field_type(Sensor, "placement")  # SensorPlacement | None
-
-    # Measurement
-    field_type(Measurement, "sensor_id")  # str
-
-    # Verify all imports are real classes (not just string names)
-    for cls in (
-        Car,
-        DiagnosticCase,
-        DrivingSegment,
-        Finding,
-        Measurement,
-        RunCapture,
-        RunSetup,
-        Sensor,
-        SensorPlacement,
-        Signature,
-        SpeedSource,
-        TestRun,
-    ):
-        assert dataclasses.is_dataclass(cls), f"{cls.__name__} must be a dataclass"
-
-    # Finding → finding-scoped value objects
-    from vibesensor.domain import (
-        ConfidenceAssessment,
-        FindingEvidence,
-        LocationHotspot,
-        VibrationOrigin,
-    )
-
-    field_type(Finding, "confidence_assessment")  # ConfidenceAssessment | None
-    field_type(Finding, "evidence")  # FindingEvidence | None
-    field_type(Finding, "location")  # LocationHotspot | None (direct field, not via origin)
-    field_type(Finding, "origin")  # VibrationOrigin | None (independent from location)
-
-    for cls in (ConfidenceAssessment, FindingEvidence, LocationHotspot, VibrationOrigin):
-        assert dataclasses.is_dataclass(cls), f"{cls.__name__} must be a dataclass"
-
-    # RunStatus is associated with Run lifecycle
-    from vibesensor.domain.run_status import RunStatus, transition_run
-
-    assert issubclass(RunStatus, str)  # StrEnum
-    assert callable(transition_run)
-
-    # DrivingPhase and DrivingSegment relationship
-    field_type(DrivingSegment, "phase")  # DrivingPhase or equivalent
-
-
-def test_finding_is_run_scoped() -> None:
-    """Finding must not reference cross-run or case-level concepts directly."""
-    import dataclasses
-
-    from vibesensor.domain import Finding
-
-    field_names = {f.name for f in dataclasses.fields(Finding)}
-    # Finding is run-scoped: it must not hold case_id, diagnosis, or
-    # cross-run aggregation fields.
-    cross_run_indicators = {"case_id", "diagnosis", "diagnoses", "test_runs", "runs", "case"}
-    leaked = field_names & cross_run_indicators
-    assert not leaked, f"Finding has cross-run fields (must be run-scoped): {leaked}"
-
-
-def test_lifecycle_mutability_rules() -> None:
-    """Run is mutable (lifecycle object); RunCapture, RunSetup, TestRun are frozen."""
-    import dataclasses
-
-    from vibesensor.domain import RunCapture, RunSetup, TestRun
-    from vibesensor.domain.run import Run
-
-    # Run is the mutable lifecycle object during recording
-    assert dataclasses.is_dataclass(Run)
-    r = Run(run_id="mut-test")
-    r.run_id = "mut-test-2"  # must not raise
-
-    # Derived/immutable objects must be frozen
-    for cls in (RunCapture, RunSetup, TestRun):
-        assert dataclasses.is_dataclass(cls), f"{cls.__name__} must be a dataclass"
-        frozen = cls.__dataclass_params__.frozen
-        assert frozen, f"{cls.__name__} must be frozen (immutable once produced)"
-
-
-# ── Domain model type completeness guardrails ───────────────────────────
-
-# Single list of domain-model types that must be importable from vibesensor.domain.
-_EXPECTED_DOMAIN_EXPORTS = [
-    # Aggregates and entities
-    "Car",
-    "DiagnosticCase",
-    "Run",
-    "TestRun",
-    # Value objects — car and context
-    "CarSnapshot",
-    "OrderReferenceSpec",
-    "TireSpec",
-    # Value objects — snapshots
-    "AnalysisSettingsSnapshot",
-    "DrivingPhaseSummary",
-    "RunContextSnapshot",
-    "RunMetadataSnapshot",
-    "SpeedProfileSummary",
-    # Value objects — run and capture
-    "ConfigurationSnapshot",
-    "Measurement",
-    "RunCapture",
-    "RunSetup",
-    "VibrationReading",
-    # Value objects — findings and diagnostics
-    "ConfidenceAssessment",
-    "Finding",
-    "FindingEvidence",
-    "FindingKind",
-    "LocationHotspot",
-    "LocationIntensitySummary",
-    "OrderMatchObservation",
-    "Signature",
-    "StrengthMetrics",
-    "StrengthPeak",
-    "VibrationOrigin",
-    "VibrationSource",
-    # Value objects — run context
-    "DrivingPhase",
-    "DrivingPhaseInterval",
-    "DrivingPhaseSegment",
-    "DrivingSegment",
-    "RunStatus",
-    "RunSuitability",
-    "Sensor",
-    "SensorPlacement",
-    "SpeedProfile",
-    "SpeedSource",
-    "SpeedSourceKind",
-    "SuitabilityCheck",
-    "Symptom",
-    # Test plan
-    "RecommendedAction",
-    "TestPlan",
-    # Functions
-    "RUN_TRANSITIONS",
-    "plan_test_actions",
-    "speed_band_sort_key",
-    "speed_bin_label",
-    "transition_run",
-]
-
-
-@pytest.mark.parametrize("name", _EXPECTED_DOMAIN_EXPORTS)
-def test_all_domain_model_types_importable(name: str) -> None:
-    """Every domain-model type must be importable from ``vibesensor.domain``."""
-    import importlib
-
-    mod = importlib.import_module("vibesensor.domain")
-    assert hasattr(mod, name), f"{name} must be exported from vibesensor.domain"
-
-
-def test_domain_exports_completeness() -> None:
-    """``__all__`` in vibesensor.domain must cover every expected domain export."""
-    import importlib
-
-    mod = importlib.import_module("vibesensor.domain")
-    all_names = set(getattr(mod, "__all__", []))
-    missing = [n for n in _EXPECTED_DOMAIN_EXPORTS if n not in all_names]
-    assert not missing, f"Missing from domain __all__: {missing}"
-
-
-# ── Car.aspects OOP: domain/use_cases must not import CarConfig ──────────
-
-
-# ---------------------------------------------------------------------------
-# Domain model migration guardrails (Step 3.10)
-# ---------------------------------------------------------------------------
-
-
-def test_no_compat_dual_base_exceptions() -> None:
-    """All VibeSensorError subclasses must not also inherit from stdlib
-    exception types (ValueError, RuntimeError, LookupError, etc.).
-
-    This prevents accidental dual-base compatibility shims that let callers
-    catch stdlib types and bypass the domain exception hierarchy.
-    """
-    from vibesensor.shared.exceptions import VibeSensorError
-
-    # stdlib exception types that should never appear as co-parents
-    stdlib_bases = (
-        ValueError,
-        TypeError,
-        RuntimeError,
-        LookupError,
-        KeyError,
-        IndexError,
-        AttributeError,
-        OSError,
-        IOError,
-        NotImplementedError,
-        ArithmeticError,
-        StopIteration,
-    )
-
-    violations: list[str] = []
-
-    def _check_recursive(cls: type) -> None:
-        # MRO between cls and VibeSensorError should contain no stdlib types
-        mro = cls.__mro__
-        vs_idx = mro.index(VibeSensorError)
-        between = mro[1:vs_idx]  # skip cls itself, stop before VibeSensorError
-        for entry in between:
-            if entry in stdlib_bases or (
-                issubclass(entry, BaseException)
-                and entry not in (VibeSensorError, Exception, BaseException)
-                and entry.__module__ == "builtins"
-            ):
-                violations.append(
-                    f"{cls.__name__} inherits from stdlib {entry.__name__} "
-                    f"(MRO: {[c.__name__ for c in mro]})"
-                )
-        for sub in cls.__subclasses__():
-            _check_recursive(sub)
-
-    _check_recursive(VibeSensorError)
-
-    assert not violations, (
-        "Custom exceptions must inherit exclusively from VibeSensorError, "
-        f"not from stdlib exception types: {violations}"
-    )
